Change reconnection interval
If the client disconnects, wait a fixed interval of 10 seconds instead of
using a larger interval at each attempt. That would make sense if the
retry attempt counter was reset to zero after each successful connection
but that's not what it was doing and the reconnection interval was quickly
growing to 5 minutes.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>